Move Harmony configuration to Harmony config isntead of EthereumJ
