+++ src/main/java/com/ethercamp/harmony/config/ApplicationConfig.java
@@ -21,7 +21,19 @@ package com.ethercamp.harmony.config;
 import com.ethercamp.harmony.service.ClientMessageService;
 import com.ethercamp.harmony.service.ClientMessageServiceDummy;
 import com.ethercamp.harmony.service.ClientMessageServiceImpl;
+import com.ethercamp.harmony.service.contracts.ContractsService;
+import com.ethercamp.harmony.service.contracts.ContractsServiceImpl;
+import com.ethercamp.harmony.service.contracts.DisabledContractService;
 import com.ethercamp.harmony.web.filter.JsonRpcUsageFilter;
+import org.apache.catalina.connector.Connector;
+import org.ethereum.config.SystemProperties;
+import org.ethereum.datasource.DbSource;
+import org.ethereum.datasource.leveldb.LevelDbDataSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
+import org.springframework.boot.context.embedded.EmbeddedServletContainerFactory;
+import org.springframework.boot.context.embedded.tomcat.TomcatEmbeddedServletContainerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Conditional;
@@ -34,6 +46,10 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 import com.ethercamp.harmony.util.exception.Web3jSafeAnnotationsErrorResolver;
+
+import java.util.ArrayList;
+import java.util.LinkedHashSet;
+import java.util.List;
 
 /**
  * Created by Stan Reshetnyk on 18.07.16.
@@ -105,6 +121,81 @@ public class ApplicationConfig extends WebMvcConfigurerAdapter {
         } else {
             return new ClientMessageServiceDummy();
         }
+    }
+
+    @Bean("contractSettingsStorage")
+    DbSource<byte[]> contractSettingsStorage() {
+        DbSource<byte[]> settingsStorage = new LevelDbDataSource("settings");
+        settingsStorage.init();
+
+        return settingsStorage;
+    }
+
+    @Bean
+    ContractsService contractsService() {
+        if (harmonyProperties(SystemProperties.getDefault()).isContractStorageEnabled()) {
+            return new ContractsServiceImpl();
+        } else {
+            return new DisabledContractService(contractSettingsStorage());
+        }
+    }
+
+    @Bean
+    HarmonyProperties harmonyProperties(SystemProperties properties) {
+        return new HarmonyProperties(properties);
+    }
+
+    @Bean
+    public EmbeddedServletContainerCustomizer containerCustomizer() {
+        return (container -> {
+            container.setPort(ports().get(0));
+        });
+    }
+
+    private List<Integer> ports() {
+        HarmonyProperties props = harmonyProperties(SystemProperties.getDefault());
+        LinkedHashSet<Integer> ports = new LinkedHashSet<>();
+
+        if (props.webPort() != null) {
+            ports.add(props.webPort());
+        }
+
+        if (props.rpcPort() != null) {
+            ports.add(props.rpcPort());
+        }
+
+        // fallback
+        if (ports.isEmpty()) {
+            ports.add(8080);
+        }
+
+        return new ArrayList<>(ports);
+    }
+
+    @Bean
+    public EmbeddedServletContainerFactory servletContainer() {
+        TomcatEmbeddedServletContainerFactory tomcat = new TomcatEmbeddedServletContainerFactory();
+        Connector[] additionalConnectors = this.additionalConnector();
+        if (additionalConnectors != null && additionalConnectors.length > 0) {
+            tomcat.addAdditionalTomcatConnectors(additionalConnectors);
+        }
+        return tomcat;
+    }
+
+    private Connector[] additionalConnector() {
+        List<Integer> ports = ports();
+        ports.remove(0);
+        if (ports().isEmpty()) {
+            return null;
+        }
+        List<Connector> result = new ArrayList<>();
+        for (Integer port : ports) {
+            Connector connector = new Connector("org.apache.coyote.http11.Http11NioProtocol");
+            connector.setScheme("http");
+            connector.setPort(port);
+            result.add(connector);
+        }
+        return result.toArray(new Connector[] {});
     }
 
     @Override
