+++ src/main/java/com/ethercamp/harmony/config/EthereumHarmonyConfig.java
@@ -53,79 +53,4 @@ import java.util.Set;
         basePackages = "org.ethereum",
         excludeFilters = @ComponentScan.Filter(NoAutoscan.class))
 public class EthereumHarmonyConfig extends CommonConfig {
-
-    @Bean
-    ContractsService contractsService() {
-        if (harmonyProperties(systemProperties()).isContractStorageEnabled()) {
-            return new ContractsServiceImpl();
-        } else {
-            return new DisabledContractService(contractSettingsStorage());
-        }
-    }
-
-    @Bean("contractSettingsStorage")
-    DbSource<byte[]> contractSettingsStorage() {
-        DbSource<byte[]> settingsStorage = new LevelDbDataSource("settings");
-        settingsStorage.init();
-
-        return settingsStorage;
-    }
-
-    @Bean
-    HarmonyProperties harmonyProperties(SystemProperties properties) {
-        return new HarmonyProperties(properties);
-    }
-
-    @Bean
-    public EmbeddedServletContainerCustomizer containerCustomizer() {
-        return (container -> {
-            container.setPort(ports().get(0));
-        });
-    }
-
-    private List<Integer> ports() {
-        HarmonyProperties props = harmonyProperties(systemProperties());
-        LinkedHashSet<Integer> ports = new LinkedHashSet<>();
-
-        if (props.webPort() != null) {
-            ports.add(props.webPort());
-        }
-
-        if (props.rpcPort() != null) {
-            ports.add(props.rpcPort());
-        }
-
-        // fallback
-        if (ports.isEmpty()) {
-            ports.add(8080);
-        }
-
-        return new ArrayList<>(ports);
-    }
-
-    @Bean
-    public EmbeddedServletContainerFactory servletContainer() {
-        TomcatEmbeddedServletContainerFactory tomcat = new TomcatEmbeddedServletContainerFactory();
-        Connector[] additionalConnectors = this.additionalConnector();
-        if (additionalConnectors != null && additionalConnectors.length > 0) {
-            tomcat.addAdditionalTomcatConnectors(additionalConnectors);
-        }
-        return tomcat;
-    }
-
-    private Connector[] additionalConnector() {
-        List<Integer> ports = ports();
-        ports.remove(0);
-        if (ports().isEmpty()) {
-            return null;
-        }
-        List<Connector> result = new ArrayList<>();
-        for (Integer port : ports) {
-            Connector connector = new Connector("org.apache.coyote.http11.Http11NioProtocol");
-            connector.setScheme("http");
-            connector.setPort(port);
-            result.add(connector);
-        }
-        return result.toArray(new Connector[] {});
-    }
 }
